refactor: move schema evolution methods to another file
